@@ -1,14 +1,22 @@
 /// Indicates the current battery state.
 enum BatteryState {
-  /// The battery is completely full of energy.
+  /// The battery is fully charged.
   full,
 
-  /// The battery is currently storing energy.
+  /// The battery is currently charging.
   charging,
+
+  /// Device is connected to external power source, but not charging the battery.
+  ///
+  /// Usually happens when device has charge limit enabled and this limit is reached.
+  /// Also, battery might be in this state if connected power source isn't powerful enough to charge the battery.
+  ///
+  /// Available on Android, MacOS and Linux platforms only.
+  connectedNotCharging,
 
   /// The battery is currently losing energy.
   discharging,
 
   /// The state of the battery is unknown.
-  unknown
+  unknown;
 }
