@@ -25,7 +25,8 @@ export 'src/model/macos_device_info.dart';
 export 'src/model/web_browser_info.dart';
 export 'src/model/windows_device_info.dart';
 
-export 'src/device_info_plus_linux.dart';
+export 'src/device_info_plus_linux.dart'
+    if (dart.library.html) 'src/device_info_plus_web.dart';
 export 'src/device_info_plus_windows.dart'
     if (dart.library.html) 'src/device_info_plus_web.dart';
 
