@@ -4,7 +4,14 @@ public struct SystemUUID {
     
     public static func getSystemUUID() -> String? {
         let dev = IOServiceMatching("IOPlatformExpertDevice")
-        let platformExpert: io_service_t = IOServiceGetMatchingService(kIOMasterPortDefault, dev)
+        
+        var platformExpert: io_service_t
+        if #available(macOS 12, *) {
+            platformExpert = IOServiceGetMatchingService(kIOMainPortDefault, dev)
+        } else {
+            platformExpert = IOServiceGetMatchingService(kIOMasterPortDefault, dev)
+        }
+        
         let serialNumberAsCFString = IORegistryEntryCreateCFProperty(platformExpert, kIOPlatformUUIDKey as CFString, kCFAllocatorDefault, 0)
         IOObjectRelease(platformExpert)
         let ser: CFTypeRef? = serialNumberAsCFString?.takeUnretainedValue()
