@@ -1,4 +1,6 @@
-const fakeAndroidBuildVersion = <String, dynamic>{
+part of '../model/android_device_info_test.dart';
+
+const _fakeAndroidBuildVersion = <String, dynamic>{
   'sdkInt': 16,
   'baseOS': 'baseOS',
   'previewSdkInt': 30,
@@ -8,19 +10,19 @@ const fakeAndroidBuildVersion = <String, dynamic>{
   'securityPatch': 'securityPatch',
 };
 
-const fakeDisplayMetrics = <String, dynamic>{
+const _fakeDisplayMetrics = <String, dynamic>{
   'widthPx': 1080.0,
   'heightPx': 2220.0,
   'xDpi': 530.0859,
   'yDpi': 529.4639,
 };
 
-const fakeSupportedAbis = <String>['arm64-v8a', 'x86', 'x86_64'];
-const fakeSupported32BitAbis = <String?>['x86 (IA-32)', 'MMX'];
-const fakeSupported64BitAbis = <String?>['x86-64', 'MMX', 'SSSE3'];
-const fakeSystemFeatures = ['FEATURE_AUDIO_PRO', 'FEATURE_AUDIO_OUTPUT'];
+const _fakeSupportedAbis = <String>['arm64-v8a', 'x86', 'x86_64'];
+const _fakeSupported32BitAbis = <String?>['x86 (IA-32)', 'MMX'];
+const _fakeSupported64BitAbis = <String?>['x86-64', 'MMX', 'SSSE3'];
+const _fakeSystemFeatures = ['FEATURE_AUDIO_PRO', 'FEATURE_AUDIO_OUTPUT'];
 
-const fakeAndroidDeviceInfo = <String, dynamic>{
+const _fakeAndroidDeviceInfo = <String, dynamic>{
   'id': 'id',
   'host': 'host',
   'tags': 'tags',
@@ -36,11 +38,11 @@ const fakeAndroidDeviceInfo = <String, dynamic>{
   'bootloader': 'bootloader',
   'fingerprint': 'fingerprint',
   'manufacturer': 'manufacturer',
-  'supportedAbis': fakeSupportedAbis,
-  'systemFeatures': fakeSystemFeatures,
-  'version': fakeAndroidBuildVersion,
-  'supported64BitAbis': fakeSupported64BitAbis,
-  'supported32BitAbis': fakeSupported32BitAbis,
-  'displayMetrics': fakeDisplayMetrics,
+  'supportedAbis': _fakeSupportedAbis,
+  'systemFeatures': _fakeSystemFeatures,
+  'version': _fakeAndroidBuildVersion,
+  'supported64BitAbis': _fakeSupported64BitAbis,
+  'supported32BitAbis': _fakeSupported32BitAbis,
+  'displayMetrics': _fakeDisplayMetrics,
   'serialNumber': 'SERIAL',
 };
