@@ -1,16 +1,16 @@
-#import "FLTHotspotNetworkInfoProvider.h"
+#import "FPPHotspotNetworkInfoProvider.h"
 #import <NetworkExtension/NetworkExtension.h>
 
-@implementation FLTHotspotNetworkInfoProvider
+@implementation FPPHotspotNetworkInfoProvider
 
 - (void)fetchNetworkInfoWithCompletionHandler:
-    (void (^)(FLTNetworkInfo *network))completionHandler
+    (void (^)(FPPNetworkInfo *network))completionHandler
     API_AVAILABLE(ios(14)) {
   [NEHotspotNetwork fetchCurrentWithCompletionHandler:^(
                         NEHotspotNetwork *network) {
     dispatch_async(dispatch_get_main_queue(), ^{
       if (network) {
-        completionHandler([[FLTNetworkInfo alloc] initWithSSID:network.SSID
+        completionHandler([[FPPNetworkInfo alloc] initWithSSID:network.SSID
                                                          BSSID:network.BSSID]);
         return;
       }
