@@ -12,7 +12,7 @@ modification, are permitted provided that the following conditions are met:
       this list of conditions and the following disclaimer in the documentation
       and/or other materials provided with the distribution.
     * The name of the author may not be used to endorse or promote products
-	  derived from this software without specific prior written permission.
+      derived from this software without specific prior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
@@ -44,6 +44,6 @@ typedef unsigned short uint16_t;
  * return value :
  *    0 : success
  *   -1 : failure    */
-/* NATPMP_LIBSPEC */int getdefaultgateway(in_addr_t * addr);
+/* NATPMP_LIBSPEC */ int getdefaultgateway(in_addr_t *addr);
 
 #endif
