@@ -4,5 +4,5 @@
 
 #import <Flutter/Flutter.h>
 
-@interface FLTPackageInfoPlusPlugin : NSObject <FlutterPlugin>
+@interface FPPPackageInfoPlusPlugin : NSObject <FlutterPlugin>
 @end
