@@ -12,7 +12,7 @@ import 'dart:io';
 import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
 
-class LANGANDCODEPAGE extends Struct {
+base class LANGANDCODEPAGE extends Struct {
   @WORD()
   external int wLanguage;
 
