@@ -1,10 +1,10 @@
 import 'dart:convert';
-import 'dart:html';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:http/http.dart';
 import 'package:package_info_plus_platform_interface/package_info_data.dart';
 import 'package:package_info_plus_platform_interface/package_info_platform_interface.dart';
+import 'package:web/helpers.dart' as web;
 
 /// The web implementation of [PackageInfoPlatform].
 ///
@@ -23,20 +23,36 @@ class PackageInfoPlusWebPlugin extends PackageInfoPlatform {
   /// Get version.json full url.
   Uri versionJsonUrl(String baseUrl, int cacheBuster) {
     final baseUri = Uri.parse(baseUrl);
-    final regExp = RegExp(r'\w+\.html.*');
-    final originPath =
+    final regExp = RegExp(r'[^/]+\.html.*');
+    final String originPath =
         '${baseUri._origin}${baseUri.path.replaceAll(regExp, '')}';
-    final versionJson = 'version.json?cachebuster=$cacheBuster';
-    return Uri.parse(originPath.endsWith('/')
-        ? '$originPath$versionJson'
-        : '$originPath/$versionJson');
+
+    // Clean uri: remove fragment and query
+    Uri uri = Uri.parse(originPath).removeFragment().replace(query: '');
+
+    // In http/s, if the last part has no file extension and not trailing slash,
+    // we can safely remove it since it's not considered as a folder
+    if (uri.path.length > 1 &&
+        !uri.path.endsWith('/') &&
+        (uri.isScheme('http') || uri.isScheme('https'))) {
+      uri = uri.replace(path: uri.path.substring(0, uri.path.lastIndexOf('/')));
+    }
+
+    // Clean uri: remove empty path segments
+    final List<String> segments = [...uri.pathSegments]
+      ..removeWhere((element) => element == '');
+
+    // Add file and cachebuster query
+    return uri.replace(
+        query: 'cachebuster=$cacheBuster',
+        pathSegments: [...segments, 'version.json']);
   }
 
   @override
   Future<PackageInfoData> getAll() async {
     final cacheBuster = DateTime.now().millisecondsSinceEpoch;
-    final url = versionJsonUrl(window.document.baseUri!, cacheBuster);
-    final response = _client == null ? await get(url) : await _client!.get(url);
+    final url = versionJsonUrl(web.window.document.baseURI, cacheBuster);
+    final response = _client == null ? await get(url) : await _client.get(url);
     final versionMap = _getVersionMap(response);
 
     return PackageInfoData(
