@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import <Flutter/Flutter.h>
+#import <FlutterMacOS/FlutterMacOS.h>
 
-@interface FLTBatteryPlusPlugin : NSObject <FlutterPlugin>
+@interface FPPPackageInfoPlusPlugin : NSObject <FlutterPlugin>
 @end
