@@ -7,50 +7,78 @@ import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.FlutterPlugin.FlutterPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.EventChannel
+import io.flutter.plugin.common.MethodChannel
 
 /** SensorsPlugin  */
 class SensorsPlugin : FlutterPlugin {
+    private lateinit var methodChannel: MethodChannel
+
     private lateinit var accelerometerChannel: EventChannel
     private lateinit var userAccelChannel: EventChannel
     private lateinit var gyroscopeChannel: EventChannel
     private lateinit var magnetometerChannel: EventChannel
 
-    private lateinit var accelerationStreamHandler: StreamHandlerImpl
-    private lateinit var linearAccelerationStreamHandler: StreamHandlerImpl
-    private lateinit var gyroScopeStreamHandler: StreamHandlerImpl
+    private lateinit var accelerometerStreamHandler: StreamHandlerImpl
+    private lateinit var userAccelStreamHandler: StreamHandlerImpl
+    private lateinit var gyroscopeStreamHandler: StreamHandlerImpl
     private lateinit var magnetometerStreamHandler: StreamHandlerImpl
 
     override fun onAttachedToEngine(binding: FlutterPluginBinding) {
+        setupMethodChannel(binding.binaryMessenger)
         setupEventChannels(binding.applicationContext, binding.binaryMessenger)
     }
 
     override fun onDetachedFromEngine(binding: FlutterPluginBinding) {
+        teardownMethodChannel()
         teardownEventChannels()
+    }
+
+    private fun setupMethodChannel(messenger: BinaryMessenger) {
+        methodChannel = MethodChannel(messenger, METHOD_CHANNEL_NAME)
+        methodChannel.setMethodCallHandler { call, result ->
+            val streamHandler = when (call.method) {
+                "setAccelerationSamplingPeriod" -> accelerometerStreamHandler
+                "setUserAccelerometerSamplingPeriod" -> userAccelStreamHandler
+                "setGyroscopeSamplingPeriod" -> gyroscopeStreamHandler
+                "setMagnetometerSamplingPeriod" -> magnetometerStreamHandler
+                else -> null
+            }
+            streamHandler?.samplingPeriod = call.arguments as Int
+            if (streamHandler != null) {
+                result.success(null)
+            } else {
+                result.notImplemented()
+            }
+        }
+    }
+
+    private fun teardownMethodChannel() {
+        methodChannel.setMethodCallHandler(null)
     }
 
     private fun setupEventChannels(context: Context, messenger: BinaryMessenger) {
         val sensorsManager = context.getSystemService(Context.SENSOR_SERVICE) as SensorManager
 
         accelerometerChannel = EventChannel(messenger, ACCELEROMETER_CHANNEL_NAME)
-        accelerationStreamHandler = StreamHandlerImpl(
+        accelerometerStreamHandler = StreamHandlerImpl(
             sensorsManager,
             Sensor.TYPE_ACCELEROMETER
         )
-        accelerometerChannel.setStreamHandler(accelerationStreamHandler)
+        accelerometerChannel.setStreamHandler(accelerometerStreamHandler)
 
         userAccelChannel = EventChannel(messenger, USER_ACCELEROMETER_CHANNEL_NAME)
-        linearAccelerationStreamHandler = StreamHandlerImpl(
+        userAccelStreamHandler = StreamHandlerImpl(
             sensorsManager,
             Sensor.TYPE_LINEAR_ACCELERATION
         )
-        userAccelChannel.setStreamHandler(linearAccelerationStreamHandler)
+        userAccelChannel.setStreamHandler(userAccelStreamHandler)
 
         gyroscopeChannel = EventChannel(messenger, GYROSCOPE_CHANNEL_NAME)
-        gyroScopeStreamHandler = StreamHandlerImpl(
+        gyroscopeStreamHandler = StreamHandlerImpl(
             sensorsManager,
             Sensor.TYPE_GYROSCOPE
         )
-        gyroscopeChannel.setStreamHandler(gyroScopeStreamHandler)
+        gyroscopeChannel.setStreamHandler(gyroscopeStreamHandler)
 
         magnetometerChannel = EventChannel(messenger, MAGNETOMETER_CHANNEL_NAME)
         magnetometerStreamHandler = StreamHandlerImpl(
@@ -66,16 +94,19 @@ class SensorsPlugin : FlutterPlugin {
         gyroscopeChannel.setStreamHandler(null)
         magnetometerChannel.setStreamHandler(null)
 
-        accelerationStreamHandler.onCancel(null)
-        linearAccelerationStreamHandler.onCancel(null)
-        gyroScopeStreamHandler.onCancel(null)
+        accelerometerStreamHandler.onCancel(null)
+        userAccelStreamHandler.onCancel(null)
+        gyroscopeStreamHandler.onCancel(null)
         magnetometerStreamHandler.onCancel(null)
     }
 
     companion object {
+        private const val METHOD_CHANNEL_NAME =
+            "dev.fluttercommunity.plus/sensors/method"
         private const val ACCELEROMETER_CHANNEL_NAME =
             "dev.fluttercommunity.plus/sensors/accelerometer"
-        private const val GYROSCOPE_CHANNEL_NAME = "dev.fluttercommunity.plus/sensors/gyroscope"
+        private const val GYROSCOPE_CHANNEL_NAME =
+            "dev.fluttercommunity.plus/sensors/gyroscope"
         private const val USER_ACCELEROMETER_CHANNEL_NAME =
             "dev.fluttercommunity.plus/sensors/user_accel"
         private const val MAGNETOMETER_CHANNEL_NAME =
