@@ -50,28 +50,32 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  static const Duration _ignoreDuration = Duration(milliseconds: 20);
+
   static const int _snakeRows = 20;
   static const int _snakeColumns = 20;
   static const double _snakeCellSize = 10.0;
 
-  List<double>? _userAccelerometerValues;
-  List<double>? _accelerometerValues;
-  List<double>? _gyroscopeValues;
-  List<double>? _magnetometerValues;
+  UserAccelerometerEvent? _userAccelerometerEvent;
+  AccelerometerEvent? _accelerometerEvent;
+  GyroscopeEvent? _gyroscopeEvent;
+  MagnetometerEvent? _magnetometerEvent;
+
+  DateTime? _userAccelerometerUpdateTime;
+  DateTime? _accelerometerUpdateTime;
+  DateTime? _gyroscopeUpdateTime;
+  DateTime? _magnetometerUpdateTime;
+
+  int? _userAccelerometerLastInterval;
+  int? _accelerometerLastInterval;
+  int? _gyroscopeLastInterval;
+  int? _magnetometerLastInterval;
   final _streamSubscriptions = <StreamSubscription<dynamic>>[];
+
+  Duration sensorInterval = SensorInterval.normalInterval;
 
   @override
   Widget build(BuildContext context) {
-    final userAccelerometer = _userAccelerometerValues
-        ?.map((double v) => v.toStringAsFixed(1))
-        .toList();
-    final accelerometer =
-        _accelerometerValues?.map((double v) => v.toStringAsFixed(1)).toList();
-    final gyroscope =
-        _gyroscopeValues?.map((double v) => v.toStringAsFixed(1)).toList();
-    final magnetometer =
-        _magnetometerValues?.map((double v) => v.toStringAsFixed(1)).toList();
-
     return Scaffold(
       appBar: AppBar(
         title: const Text('Sensors Plus Example'),
@@ -97,40 +101,118 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
           ),
           Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Text('UserAccelerometer: $userAccelerometer'),
+            padding: const EdgeInsets.all(20.0),
+            child: Table(
+              columnWidths: const {
+                0: FlexColumnWidth(4),
+                4: FlexColumnWidth(2),
+              },
+              children: [
+                const TableRow(
+                  children: [
+                    SizedBox.shrink(),
+                    Text('X'),
+                    Text('Y'),
+                    Text('Z'),
+                    Text('Interval'),
+                  ],
+                ),
+                TableRow(
+                  children: [
+                    const Padding(
+                      padding: EdgeInsets.symmetric(vertical: 8.0),
+                      child: Text('UserAccelerometer'),
+                    ),
+                    Text(_userAccelerometerEvent?.x.toStringAsFixed(1) ?? '?'),
+                    Text(_userAccelerometerEvent?.y.toStringAsFixed(1) ?? '?'),
+                    Text(_userAccelerometerEvent?.z.toStringAsFixed(1) ?? '?'),
+                    Text(
+                        '${_userAccelerometerLastInterval?.toString() ?? '?'} ms'),
+                  ],
+                ),
+                TableRow(
+                  children: [
+                    const Padding(
+                      padding: EdgeInsets.symmetric(vertical: 8.0),
+                      child: Text('Accelerometer'),
+                    ),
+                    Text(_accelerometerEvent?.x.toStringAsFixed(1) ?? '?'),
+                    Text(_accelerometerEvent?.y.toStringAsFixed(1) ?? '?'),
+                    Text(_accelerometerEvent?.z.toStringAsFixed(1) ?? '?'),
+                    Text('${_accelerometerLastInterval?.toString() ?? '?'} ms'),
+                  ],
+                ),
+                TableRow(
+                  children: [
+                    const Padding(
+                      padding: EdgeInsets.symmetric(vertical: 8.0),
+                      child: Text('Gyroscope'),
+                    ),
+                    Text(_gyroscopeEvent?.x.toStringAsFixed(1) ?? '?'),
+                    Text(_gyroscopeEvent?.y.toStringAsFixed(1) ?? '?'),
+                    Text(_gyroscopeEvent?.z.toStringAsFixed(1) ?? '?'),
+                    Text('${_gyroscopeLastInterval?.toString() ?? '?'} ms'),
+                  ],
+                ),
+                TableRow(
+                  children: [
+                    const Padding(
+                      padding: EdgeInsets.symmetric(vertical: 8.0),
+                      child: Text('Magnetometer'),
+                    ),
+                    Text(_magnetometerEvent?.x.toStringAsFixed(1) ?? '?'),
+                    Text(_magnetometerEvent?.y.toStringAsFixed(1) ?? '?'),
+                    Text(_magnetometerEvent?.z.toStringAsFixed(1) ?? '?'),
+                    Text('${_magnetometerLastInterval?.toString() ?? '?'} ms'),
+                  ],
+                ),
               ],
             ),
           ),
-          Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Text('Accelerometer: $accelerometer'),
-              ],
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Text('Gyroscope: $gyroscope'),
-              ],
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Text('Magnetometer: $magnetometer'),
-              ],
-            ),
+          Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              const Text('Update Interval:'),
+              SegmentedButton(
+                segments: [
+                  ButtonSegment(
+                    value: SensorInterval.gameInterval,
+                    label: Text('Game\n'
+                        '(${SensorInterval.gameInterval.inMilliseconds}ms)'),
+                  ),
+                  ButtonSegment(
+                    value: SensorInterval.uiInterval,
+                    label: Text('UI\n'
+                        '(${SensorInterval.uiInterval.inMilliseconds}ms)'),
+                  ),
+                  ButtonSegment(
+                    value: SensorInterval.normalInterval,
+                    label: Text('Normal\n'
+                        '(${SensorInterval.normalInterval.inMilliseconds}ms)'),
+                  ),
+                  const ButtonSegment(
+                    value: Duration(milliseconds: 500),
+                    label: Text('500ms'),
+                  ),
+                  const ButtonSegment(
+                    value: Duration(seconds: 1),
+                    label: Text('1s'),
+                  ),
+                ],
+                selected: {sensorInterval},
+                showSelectedIcon: false,
+                onSelectionChanged: (value) {
+                  setState(() {
+                    sensorInterval = value.first;
+                    userAccelerometerEventStream(
+                        samplingPeriod: sensorInterval);
+                    accelerometerEventStream(samplingPeriod: sensorInterval);
+                    gyroscopeEventStream(samplingPeriod: sensorInterval);
+                    magnetometerEventStream(samplingPeriod: sensorInterval);
+                  });
+                },
+              ),
+            ],
           ),
         ],
       ),
@@ -149,53 +231,19 @@ class _MyHomePageState extends State<MyHomePage> {
   void initState() {
     super.initState();
     _streamSubscriptions.add(
-      userAccelerometerEvents.listen(
+      userAccelerometerEventStream(samplingPeriod: sensorInterval).listen(
         (UserAccelerometerEvent event) {
+          final now = DateTime.now();
           setState(() {
-            _userAccelerometerValues = <double>[event.x, event.y, event.z];
+            _userAccelerometerEvent = event;
+            if (_userAccelerometerUpdateTime != null) {
+              final interval = now.difference(_userAccelerometerUpdateTime!);
+              if (interval > _ignoreDuration) {
+                _userAccelerometerLastInterval = interval.inMilliseconds;
+              }
+            }
           });
-        },
-        onError: (e) {
-          showDialog(
-              context: context,
-              builder: (context) {
-                return const AlertDialog(
-                  title: Text("Sensor Not Found"),
-                  content: Text(
-                      "It seems that your device doesn't support Accelerometer Sensor"),
-                );
-              });
-        },
-        cancelOnError: true,
-      ),
-    );
-    _streamSubscriptions.add(
-      accelerometerEvents.listen(
-        (AccelerometerEvent event) {
-          setState(() {
-            _accelerometerValues = <double>[event.x, event.y, event.z];
-          });
-        },
-        onError: (e) {
-          showDialog(
-              context: context,
-              builder: (context) {
-                return const AlertDialog(
-                  title: Text("Sensor Not Found"),
-                  content: Text(
-                      "It seems that your device doesn't support Gyroscope Sensor"),
-                );
-              });
-        },
-        cancelOnError: true,
-      ),
-    );
-    _streamSubscriptions.add(
-      gyroscopeEvents.listen(
-        (GyroscopeEvent event) {
-          setState(() {
-            _gyroscopeValues = <double>[event.x, event.y, event.z];
-          });
+          _userAccelerometerUpdateTime = now;
         },
         onError: (e) {
           showDialog(
@@ -212,11 +260,77 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
     );
     _streamSubscriptions.add(
-      magnetometerEvents.listen(
-        (MagnetometerEvent event) {
+      accelerometerEventStream(samplingPeriod: sensorInterval).listen(
+        (AccelerometerEvent event) {
+          final now = DateTime.now();
           setState(() {
-            _magnetometerValues = <double>[event.x, event.y, event.z];
+            _accelerometerEvent = event;
+            if (_accelerometerUpdateTime != null) {
+              final interval = now.difference(_accelerometerUpdateTime!);
+              if (interval > _ignoreDuration) {
+                _accelerometerLastInterval = interval.inMilliseconds;
+              }
+            }
           });
+          _accelerometerUpdateTime = now;
+        },
+        onError: (e) {
+          showDialog(
+              context: context,
+              builder: (context) {
+                return const AlertDialog(
+                  title: Text("Sensor Not Found"),
+                  content: Text(
+                      "It seems that your device doesn't support Accelerometer Sensor"),
+                );
+              });
+        },
+        cancelOnError: true,
+      ),
+    );
+    _streamSubscriptions.add(
+      gyroscopeEventStream(samplingPeriod: sensorInterval).listen(
+        (GyroscopeEvent event) {
+          final now = DateTime.now();
+          setState(() {
+            _gyroscopeEvent = event;
+            if (_gyroscopeUpdateTime != null) {
+              final interval = now.difference(_gyroscopeUpdateTime!);
+              if (interval > _ignoreDuration) {
+                _gyroscopeLastInterval = interval.inMilliseconds;
+              }
+            }
+          });
+          _gyroscopeUpdateTime = now;
+        },
+        onError: (e) {
+          showDialog(
+              context: context,
+              builder: (context) {
+                return const AlertDialog(
+                  title: Text("Sensor Not Found"),
+                  content: Text(
+                      "It seems that your device doesn't support Gyroscope Sensor"),
+                );
+              });
+        },
+        cancelOnError: true,
+      ),
+    );
+    _streamSubscriptions.add(
+      magnetometerEventStream(samplingPeriod: sensorInterval).listen(
+        (MagnetometerEvent event) {
+          final now = DateTime.now();
+          setState(() {
+            _magnetometerEvent = event;
+            if (_magnetometerUpdateTime != null) {
+              final interval = now.difference(_magnetometerUpdateTime!);
+              if (interval > _ignoreDuration) {
+                _magnetometerLastInterval = interval.inMilliseconds;
+              }
+            }
+          });
+          _magnetometerUpdateTime = now;
         },
         onError: (e) {
           showDialog(
