@@ -84,7 +84,7 @@ class SnakeState extends State<Snake> {
   void initState() {
     super.initState();
     _streamSubscription =
-        accelerometerEvents.listen((AccelerometerEvent event) {
+        accelerometerEventStream().listen((AccelerometerEvent event) {
       setState(() {
         acceleration = event;
       });
