@@ -55,8 +55,11 @@ class WebSensorsPlugin extends SensorsPlatform {
   StreamController<AccelerometerEvent>? _accelerometerStreamController;
   late Stream<AccelerometerEvent> _accelerometerResultStream;
 
+  // todo: make web also support setting samplingPeriod
   @override
-  Stream<AccelerometerEvent> get accelerometerEvents {
+  Stream<AccelerometerEvent> accelerometerEventStream({
+    Duration samplingPeriod = SensorInterval.normalInterval,
+  }) {
     if (_accelerometerStreamController == null) {
       _accelerometerStreamController = StreamController<AccelerometerEvent>();
       _featureDetected(
@@ -95,6 +98,10 @@ class WebSensorsPlugin extends SensorsPlatform {
       );
       _accelerometerResultStream =
           _accelerometerStreamController!.stream.asBroadcastStream();
+
+      _accelerometerStreamController!.onCancel = () {
+        _accelerometerStreamController!.close();
+      };
     }
 
     return _accelerometerResultStream;
@@ -103,8 +110,11 @@ class WebSensorsPlugin extends SensorsPlatform {
   StreamController<GyroscopeEvent>? _gyroscopeEventStreamController;
   late Stream<GyroscopeEvent> _gyroscopeEventResultStream;
 
+  // todo: make web also support setting samplingPeriod
   @override
-  Stream<GyroscopeEvent> get gyroscopeEvents {
+  Stream<GyroscopeEvent> gyroscopeEventStream({
+    Duration samplingPeriod = SensorInterval.normalInterval,
+  }) {
     if (_gyroscopeEventStreamController == null) {
       _gyroscopeEventStreamController = StreamController<GyroscopeEvent>();
       _featureDetected(
@@ -155,8 +165,11 @@ class WebSensorsPlugin extends SensorsPlatform {
   StreamController<UserAccelerometerEvent>? _userAccelerometerStreamController;
   late Stream<UserAccelerometerEvent> _userAccelerometerResultStream;
 
+  // todo: make web also support setting samplingPeriod
   @override
-  Stream<UserAccelerometerEvent> get userAccelerometerEvents {
+  Stream<UserAccelerometerEvent> userAccelerometerEventStream({
+    Duration samplingPeriod = SensorInterval.normalInterval,
+  }) {
     if (_userAccelerometerStreamController == null) {
       _userAccelerometerStreamController =
           StreamController<UserAccelerometerEvent>();
@@ -197,6 +210,10 @@ class WebSensorsPlugin extends SensorsPlatform {
       );
       _userAccelerometerResultStream =
           _userAccelerometerStreamController!.stream.asBroadcastStream();
+
+      _userAccelerometerStreamController!.onCancel = () {
+        _userAccelerometerStreamController!.close();
+      };
     }
 
     return _userAccelerometerResultStream;
@@ -205,8 +222,11 @@ class WebSensorsPlugin extends SensorsPlatform {
   StreamController<MagnetometerEvent>? _magnetometerStreamController;
   late Stream<MagnetometerEvent> _magnetometerResultStream;
 
+  // todo: make web also support setting samplingPeriod
   @override
-  Stream<MagnetometerEvent> get magnetometerEvents {
+  Stream<MagnetometerEvent> magnetometerEventStream({
+    Duration samplingPeriod = SensorInterval.normalInterval,
+  }) {
     if (_magnetometerStreamController == null) {
       _magnetometerStreamController = StreamController<MagnetometerEvent>();
       _featureDetected(
@@ -245,6 +265,10 @@ class WebSensorsPlugin extends SensorsPlatform {
       );
       _magnetometerResultStream =
           _magnetometerStreamController!.stream.asBroadcastStream();
+
+      _magnetometerStreamController!.onCancel = () {
+        _magnetometerStreamController!.close();
+      };
     }
 
     return _magnetometerResultStream;
